+++ apps/app/src/routes/_layout/profile/_tabs/overview.tsx
@@ -1,10 +1,26 @@
+import { ApproverFor } from "@/components/profile/overview/ApproverFor";
 import { createFileRoute } from "@tanstack/react-router";
-import { ProfileOverview } from "../../../../components/profile/overview";
+// import { CurateCTA } from "./CurateCTA";
+import { CuratorFor } from "@/components/profile/overview/CuratorFor";
+// import { TopBadges } from "./TopBadges";
+// import { UserStats } from "./UserStats";
 
 export const Route = createFileRoute("/_layout/profile/_tabs/overview")({
-  component: RouteComponent,
+  component: ProfileOverview,
 });
 
-function RouteComponent() {
-  return <ProfileOverview />;
+function ProfileOverview() {
+  return (
+    <div className="flex flex-col items-stretch gap-4 sm:gap-6">
+      {/* <div className="grid grid-cols-1 md:grid-cols-2 gap-4 sm:gap-6">
+        <UserStats />
+        <TopBadges />
+      </div> */}
+      {/* <CurateCTA /> */}
+      <div className="grid grid-cols-1 md:grid-cols-2 gap-4 sm:gap-6">
+        <ApproverFor />
+        <CuratorFor />
+      </div>
+    </div>
+  );
 }
