+++ apps/app/src/routes/_layout/profile/_tabs/settings/_tabs/notifications.tsx
@@ -1,11 +1,11 @@
 import { createFileRoute } from "@tanstack/react-router";
-import { Card } from "../../../../components/ui/card";
+import { Card } from "../../../../../../components/ui/card";
 
-export const Route = createFileRoute("/_layout/profile/settings/notifications")(
-  {
-    component: NotificationsComponent,
-  },
-);
+export const Route = createFileRoute(
+  "/_layout/profile/_tabs/settings/_tabs/notifications",
+)({
+  component: NotificationsComponent,
+});
 
 function NotificationsComponent() {
   return (
