+++ apps/app/src/routes/_layout/profile/index.tsx
@@ -2,6 +2,6 @@ import { createFileRoute, redirect } from "@tanstack/react-router";
 
 export const Route = createFileRoute("/_layout/profile/")({
   beforeLoad: () => {
-    throw redirect({ to: "/profile/overview" });
+    throw redirect({ to: "/profile/overview", search: {} });
   },
 });
