--- apps/app/src/routes/_layout/profile/settings/index.tsx
@@ -1,5 +1,0 @@
-import { createFileRoute, Navigate } from "@tanstack/react-router";
-
-export const Route = createFileRoute("/_layout/profile/settings/")({
-  component: () => <Navigate to="/profile/settings/connections" />,
-});
